api_todo/requirements.txt
asgiref==3.3.1
Django==3.1.4
djangorestframework==3.12.2
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.4
sqlparse==0.4.1
